serverless/requirements.txt
pyjwkest>=1.4.2
requests>=2.22.0
cryptography>=2.8
boto3>=1.9.37
